Removed (size_t) casts done for Boost 1.35.
Boost's iterator_range::size() returns difference_type instead of
size_type in Boost 1.35.  I guess that makes some sense in the context
of a general iterator_range, but for my use as a substring range, it's
annoying.  However, I'm no longer using Boost for that.  There's a new
class, StringRange.  So I removed the hacks to make to code work with
Boost.

+++ src/cc/warp/bloom_filter.cc
@@ -111,7 +111,7 @@ void BloomFilter::load(strref_t buf)
     size_t nSeeds = deserialize<uint32_t>(p);
     p += 4;
 
-    if((size_t)buf.size() != (4 + 4 + (4 * nSeeds) + ((nBits + 7) / 8)))
+    if(buf.size() != (4 + 4 + (4 * nSeeds) + ((nBits + 7) / 8)))
         raise<ValueError>("Bloom filter data is wrong size");
 
     seeds.clear();
